requirements.txt
# Emotion AI Application - Dependencies
# Stack: Streamlit + SQLite + Claude API
# Version légère sans TensorFlow/PyTorch

# Web Framework
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Authentication
bcrypt>=4.0.0
pyjwt>=2.8.0

# Computer Vision (léger - OpenCV uniquement)
opencv-python-headless>=4.8.0
numpy>=1.24.0

# AI Conversational
groq>=0.4.0              # 🆓 GRATUIT - Recommandé
anthropic>=0.18.0        # 💰 Payant - Optionnel

# Data Visualization
plotly>=5.18.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# Webcam pour Streamlit
streamlit-webrtc>=0.47.0
av>=10.0.0

# ============================================
# OPTIONNEL: Pour une meilleure détection d'émotions
# Décommentez les lignes ci-dessous si vous avez une bonne connexion
# ============================================
# tensorflow>=2.13.0
# fer>=22.5.1

